TP2/api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-telegram-bot==13.7
paho-mqtt==1.6.1